printProgress(): Remove useless casting
Java's "binary numeric promotion" applies here:
One of the operands is a double, the other is a float, so the float is
automatically converted to double.

0d is also unnecessary due to "Widening Primitive Conversion".

+++ src/main/java/checkpoint/generation/ConcurrentCheckpointGenerator.java
@@ -470,7 +470,7 @@ public final class ConcurrentCheckpointGenerator
 		
 		// Need to use double here because byte counts are easily billions.
 		double bytesPerSec = elapsedSecs > 0
-			? (double)finishedBytes / (double)elapsedSecs : 0d;
+			? (double)finishedBytes / elapsedSecs : 0;
 		double mibPerSec   = Math.scalb(bytesPerSec, -20);
 		
 		String remainingTime;
